split BlinkImpl into multiple files
