+++ extensions/warp-blink-wrtc/src/blink_impl/call_initiation.rs
@@ -1,0 +1,139 @@
+use crate::host_media;
+use crate::simple_webrtc;
+
+use async_trait::async_trait;
+use host_media::{
+    audio::{
+        automute::{AutoMuteCmd, AUDIO_CMD_CH},
+        AudioCodec, AudioHardwareConfig,
+    },
+    mp4_logger::Mp4LoggerConfig,
+};
+use std::{
+    any::Any,
+    collections::{HashMap, HashSet},
+    str::FromStr,
+    sync::Arc,
+    time::Duration,
+};
+use webrtc::rtp_transceiver::rtp_codec::RTCRtpCodecCapability;
+
+use anyhow::{bail, Context};
+use cpal::traits::{DeviceTrait, HostTrait};
+use futures::StreamExt;
+
+use rust_ipfs::{Ipfs, Keypair, SubscriptionStream};
+
+use tokio::{
+    sync::{
+        broadcast::{self, Sender},
+        RwLock,
+    },
+    task::JoinHandle,
+};
+use uuid::Uuid;
+use warp::{
+    blink::{AudioDeviceConfig, Blink, BlinkEventKind, BlinkEventStream, CallConfig, CallInfo},
+    crypto::{did_key::Generate, zeroize::Zeroizing, DIDKey, Ed25519KeyPair, Fingerprint, DID},
+    error::Error,
+    module::Module,
+    multipass::MultiPass,
+    Extension, SingleHandle,
+};
+
+use crate::{
+    host_media::audio::AudioSampleRate,
+    signaling::{ipfs_routes, CallSignal, InitiationSignal, PeerSignal},
+    simple_webrtc::events::{EmittedEvents, WebRtcEventStream},
+    store::{
+        decode_gossipsub_msg_aes, decode_gossipsub_msg_ecdh, send_signal_aes, send_signal_ecdh,
+        PeerIdExt,
+    },
+};
+
+use super::data::PendingCall;
+
+pub async fn run(
+    own_id: Arc<RwLock<Option<DID>>>,
+    pending_calls: Arc<RwLock<HashMap<Uuid, PendingCall>>>,
+    mut stream: SubscriptionStream,
+    ch: Sender<BlinkEventKind>,
+) {
+    while let Some(msg) = stream.next().await {
+        let sender = match msg.source.and_then(|s| s.to_did().ok()) {
+            Some(id) => id,
+            None => {
+                log::error!("msg received without source");
+                continue;
+            }
+        };
+
+        let signal: InitiationSignal = {
+            let lock = own_id.read().await;
+            let own_id = match lock.as_ref().ok_or(Error::BlinkNotInitialized) {
+                Ok(r) => r,
+                Err(e) => {
+                    log::error!("{e}");
+                    continue;
+                }
+            };
+
+            match decode_gossipsub_msg_ecdh(own_id, &sender, &msg) {
+                Ok(s) => s,
+                Err(e) => {
+                    log::error!("failed to decode msg from call initiation stream: {e}");
+                    continue;
+                }
+            }
+        };
+
+        match signal {
+            InitiationSignal::Offer { call_info } => {
+                if !call_info.participants().contains(&sender) {
+                    log::warn!("someone offered a call for which they weren't a participant");
+                    continue;
+                }
+                let call_id = call_info.call_id();
+                let evt = BlinkEventKind::IncomingCall {
+                    call_id,
+                    conversation_id: call_info.conversation_id(),
+                    sender: sender.clone(),
+                    participants: call_info.participants(),
+                };
+
+                let pc = PendingCall {
+                    call: call_info,
+                    connected_participants: HashSet::from_iter(vec![sender].drain(..)),
+                };
+                pending_calls.write().await.insert(call_id, pc);
+                if let Err(e) = ch.send(evt) {
+                    log::error!("failed to send IncomingCall event: {e}");
+                }
+            }
+            InitiationSignal::Join { call_id } => {
+                if let Some(pc) = pending_calls.write().await.get_mut(&call_id) {
+                    if !pc.call.participants().contains(&sender) {
+                        log::warn!("someone who wasn't a participant tried to cancel the call");
+                        continue;
+                    }
+                    pc.connected_participants.insert(sender);
+                }
+            }
+            InitiationSignal::Leave { call_id } => {
+                if let Some(pc) = pending_calls.write().await.get_mut(&call_id) {
+                    if !pc.call.participants().contains(&sender) {
+                        log::warn!("someone who wasn't a participant tried to cancel the call");
+                        continue;
+                    }
+                    pc.connected_participants.remove(&sender);
+                    if pc.connected_participants.is_empty() {
+                        let evt = BlinkEventKind::CallCancelled { call_id };
+                        if let Err(e) = ch.send(evt) {
+                            log::error!("failed to send CallCancelled event: {e}");
+                        }
+                    }
+                }
+            }
+        }
+    }
+}
